converted ints to shorts to safe space

+++ src/de/alex/byteSerializer/Buffer.java
@@ -6,7 +6,7 @@ public class Buffer {
 	public static void putSizeIn(int startPos,ByteBuffer buffer){
 		int endPosition = buffer.position();
 		buffer.position(startPos);
-		buffer.putInt(endPosition-startPos);
+		buffer.putShort((short) (endPosition-startPos));
 		buffer.position(endPosition);
 	}
 }
