+++ src/de/alex/byteSerializer/ByteSerializer.java
@@ -6,16 +6,16 @@ public class ByteSerializer extends TypeSerializer{
 	@Override
 	protected void serialize(Object object, ByteBuffer buffer) {
 		int startPosition = buffer.position();
-		buffer.putInt(Integer.MAX_VALUE); //as a placeholder
-		buffer.putInt(((byte[]) object).length);//useless
+		buffer.putShort(Short.MAX_VALUE); //as a placeholder
+		buffer.putShort(((short) ((byte[]) object).length));//useless
 		buffer.put(((byte[]) object));
 		Buffer.putSizeIn(startPosition,buffer);
 	}
 
 	@Override
 	protected Object deserialize(ByteBuffer buffer) {
-		int size = buffer.getInt();
-		int arraySize = buffer.getInt();
+		int size = (int)buffer.getShort();
+		int arraySize = buffer.getShort();
 		byte[] out = new byte[arraySize];
 		buffer.get(out);
 		return out;
