+++ src/de/alex/byteSerializer/StringSerializer.java
@@ -8,13 +8,13 @@ public class StringSerializer extends TypeSerializer{
 	@Override
 	protected void serialize(Object object, ByteBuffer buffer) {
 		byte[] bytes = ((String) object).getBytes();
-		buffer.putInt(bytes.length);
+		buffer.putShort((short) bytes.length);
 		buffer.put(bytes);
 	}
 
 	@Override
 	protected Object deserialize(ByteBuffer buffer) {
-		int size = buffer.getInt();
+		int size = buffer.getShort();
 		byte[] outString = new byte[size];
 		buffer.get(outString);
 		return new String(outString);
